inline helper function for comparing floats
