+++ .changeset/quick-cobras-vanish.md
@@ -1,0 +1,5 @@
+---
+"@rnx-kit/metro-service": patch
+---
+
+Inline helper function for comparing floats
