+++ packages/metro-service/test/asset/android.test.ts
@@ -1,5 +1,8 @@
 import type { PackagerAsset } from "../../src/asset/types";
-import { getAndroidAssetSuffix } from "../../src/asset/android";
+import {
+  getAndroidAssetSuffix,
+  isApproximatelyEqual,
+} from "../../src/asset/android";
 
 describe("MetroService", () => {
   const asset: PackagerAsset = {
@@ -50,5 +53,21 @@ describe("MetroService", () => {
     expect(() =>
       getAndroidAssetSuffix(asset, 1.23)
     ).toThrowErrorMatchingSnapshot();
+  });
+
+  test("isApproximatelyEqual(1.25, 1.2499, tolerance=0.001) returns true", () => {
+    expect(isApproximatelyEqual(1.25, 1.2499, 0.001)).toBe(true);
+  });
+
+  test("isApproximatelyEqual(1.25, 1.2499, tolerance=0.000001) returns false", () => {
+    expect(isApproximatelyEqual(1.25, 1.2499, 0.000001)).toBe(false);
+  });
+
+  test("isApproximatelyEqual(10, 11, tolerance=2) returns true", () => {
+    expect(isApproximatelyEqual(10, 11, 2)).toBe(true);
+  });
+
+  test("isApproximatelyEqual(10, 11, tolerance=0.5) returns false", () => {
+    expect(isApproximatelyEqual(10, 11, 0.5)).toBe(false);
   });
 });
