--- packages/tools-language/math.d.ts
@@ -1,1 +1,0 @@
-export * from "./lib/math";
