--- packages/tools-language/math.js
@@ -1,1 +1,0 @@
-module.exports = require("./lib/math");
