+++ packages/tools-language/src/index.ts
@@ -1,3 +1,2 @@
 export { isNonEmptyArray } from "./array";
-export { isApproximatelyEqual } from "./math";
 export { hasProperty, keysOf, pickValues } from "./properties";
